+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## [Unreleased]
 
 
+## [v347] - 2026-02-02
+
+- JRuby 10.0.3.0 is now available
+
+
+## [v346] - 2026-02-02
+
+- Add metric: Track apps with and without `BUNDLED WITH` in the `Gemfile.lock`
+- Default bundler version changed from `2.3.25` to `2.5.23`. Ruby's [standard version of bundler](https://stdgems.org/bundler/) will take precedence if it is greater than Heroku's installed version. When there is no `BUNDLED WITH` in the `Gemfile.lock`, then `bundle install` will use the highest version of Bundler available.
+
 ## [v345] - 2026-01-26
 
 - Set `WEB_CONCURRENCY_SET_BY=heroku/ruby` if `WEB_CONCURRENCY` gets set in (deprecated) `SENSIBLE_DEFAULTS` mode (https://github.com/heroku/heroku-buildpack-ruby/pull/1700)
@@ -1894,7 +1904,9 @@ Bugfixes:
 * Change gem detection to use lockfile parser
 * use `$RACK_ENV` when thin is detected for rack apps
 
-[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v345...main
+[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v347...main
+[v347]: https://github.com/heroku/heroku-buildpack-ruby/compare/v346...v347
+[v346]: https://github.com/heroku/heroku-buildpack-ruby/compare/v345...v346
 [v345]: https://github.com/heroku/heroku-buildpack-ruby/compare/v344...v345
 [v344]: https://github.com/heroku/heroku-buildpack-ruby/compare/v343...v344
 [v343]: https://github.com/heroku/heroku-buildpack-ruby/compare/v342...v343